Refactor cloudflare-rs cli to call from http.rs
